Refactor add-on toolbar button & panel into module
- Create `addon/lib/toolbar-button.js` module to organize code related
  to the toolbar button and the panel it pops up

- Move code from `addon/index.js` to the new module

Issue #610
